网络请求封装

+++ WinLottery/third part/Classes/CycleScrollView.m
@@ -140,7 +140,7 @@
     }
     if(contentOffsetX <= 0) {
         self.currentPageIndex = [self getValidNextPageIndexWithPageIndex:self.currentPageIndex - 1];
-        NSLog(@"previous，当前页:%d",self.currentPageIndex);
+        NSLog(@"previous，当前页:%ld",(long)self.currentPageIndex);
         [self configContentViews];
     }
 }
